GetTFTargetGenes

Retrieve target genes for a set of transcription factors (TFs) from a Seurat object, based on either regulons or the full TF network. The function allows exploration of direct targets or extending the network by adding target genes of TFs that are downstream in the network.

Usage

GetTFTargetGenes(
  seurat_obj,
  selected_tfs,
  depth = 1,
  target_type = "both",
  use_regulons = TRUE,
  wgcna_name = NULL
)

Arguments

seurat_obj

A Seurat object containing the TF network or regulon information. This must include WGCNA results in the @misc slot.

selected_tfs

A list of transcription factors (TFs) to use as the starting point for network exploration.

depth

Integer value specifying the number of layers to extend the TF network from the selected TFs. For example, if depth=2, the target genes of selected_tfs are shown, and additional target genes are shown for TFs that are targets of the original selected_tfs.

target_type

The type of target genes to include in the output. Options are "positive" (genes with expression positively correlated with TFs), "negative" (genes with expression negatively correlated with TFs), or "both" (default, includes all targets).

use_regulons

Logical flag indicating whether to use regulons (default = TRUE) or the entire TF network for target gene discovery.

wgcna_name

Character string specifying the name of the WGCNA experiment. This should be stored in the @misc slot of the Seurat object.

Value

A data frame containing the TF-target interactions at each specified depth level, with additional information such as regulatory score, correlation, and depth.

Details

This function retrieves direct and extended targets of a set of TFs based on the regulatory network stored in the Seurat object. The depth parameter controls how many layers of TF-target interactions to explore, while the target_type parameter allows filtering based on correlation direction. The use_regulons flag specifies whether to use regulon data, which may provide a more confident set of interactions, or the full TF network.