ModuleNetworkPlot

Visualizes the top hub genes for selected modules as a circular network plot

ModuleNetworkPlot(
  seurat_obj,
  mods = "all",
  outdir = "ModuleNetworks",
  plot_size = c(6, 6),
  wgcna_name = NULL,
  label_center = FALSE,
  edge.alpha = 0.25,
  vertex.label.cex = 1,
  vertex.size = 6,
  ...
)

Arguments

seurat_obj

A Seurat object

mods

Names of the modules to plot. If mods = "all", all modules are plotted.

outdir

The directory where the plots will be stored.

plot_size

A vector containing the width and height of the network plots.

wgcna_name

The name of the hdWGCNA experiment in the seurat_obj@misc slot

Examples

ModuleNetworkPlot
#> function (seurat_obj, mods = "all", outdir = "ModuleNetworks", 
#>     plot_size = c(6, 6), wgcna_name = NULL, label_center = FALSE, 
#>     edge.alpha = 0.25, vertex.label.cex = 1, vertex.size = 6, 
#>     ...) 
#> {
#>     if (is.null(wgcna_name)) {
#>         wgcna_name <- seurat_obj@misc$active_wgcna
#>     }
#>     MEs <- GetMEs(seurat_obj, wgcna_name)
#>     modules <- GetModules(seurat_obj, wgcna_name)
#>     if (mods == "all") {
#>         mods <- levels(modules$module)
#>         mods <- mods[mods != "grey"]
#>     }
#>     if (!dir.exists(outdir)) {
#>         dir.create(outdir)
#>     }
#>     cat(paste0("Writing output files to ", outdir))
#>     TOM <- GetTOM(seurat_obj, wgcna_name)
#>     n_hubs <- 25
#>     hub_list <- lapply(mods, function(cur_mod) {
#>         cur <- subset(modules, module == cur_mod)
#>         cur <- cur[, c("gene_name", paste0("kME_", cur_mod))] %>% 
#>             top_n(n_hubs)
#>         colnames(cur)[2] <- "var"
#>         cur %>% arrange(desc(var)) %>% .$gene_name
#>     })
#>     names(hub_list) <- mods
#>     for (cur_mod in mods) {
#>         print(cur_mod)
#>         cur_color <- modules %>% subset(module == cur_mod) %>% 
#>             .$color %>% unique
#>         n_genes = 25
#>         n_conns = 500
#>         cur_kME <- paste0("kME_", cur_mod)
#>         cur_genes <- hub_list[[cur_mod]]
#>         matchind <- match(cur_genes, colnames(TOM))
#>         reducedTOM = TOM[matchind, matchind]
#>         orderind <- order(reducedTOM, decreasing = TRUE)
#>         connections2keep <- orderind[1:n_conns]
#>         reducedTOM <- matrix(0, nrow(reducedTOM), ncol(reducedTOM))
#>         reducedTOM[connections2keep] <- 1
#>         if (label_center) {
#>             cur_genes[11:25] <- ""
#>         }
#>         gA <- graph.adjacency(as.matrix(reducedTOM[1:10, 1:10]), 
#>             mode = "undirected", weighted = TRUE, diag = FALSE)
#>         gB <- graph.adjacency(as.matrix(reducedTOM[11:n_genes, 
#>             11:n_genes]), mode = "undirected", weighted = TRUE, 
#>             diag = FALSE)
#>         layoutCircle <- rbind(layout.circle(gA)/2, layout.circle(gB))
#>         g1 <- graph.adjacency(as.matrix(reducedTOM), mode = "undirected", 
#>             weighted = TRUE, diag = FALSE)
#>         pdf(paste0(outdir, "/", cur_mod, ".pdf"), width = plot_size[1], 
#>             height = plot_size[2], useDingbats = FALSE)
#>         plot(g1, edge.color = adjustcolor(cur_color, alpha.f = 0.25), 
#>             edge.alpha = edge.alpha, vertex.color = cur_color, 
#>             vertex.label = as.character(cur_genes), vertex.label.dist = 1.1, 
#>             vertex.label.degree = -pi/4, vertex.label.color = "black", 
#>             vertex.label.family = "Helvetica", vertex.label.font = 3, 
#>             vertex.label.cex = vertex.label.cex, vertex.frame.color = "black", 
#>             layout = jitter(layoutCircle), vertex.size = vertex.size, 
#>             main = paste(cur_mod))
#>         dev.off()
#>     }
#> }
#> <bytecode: 0x7fe5929a4de8>
#> <environment: namespace:hdWGCNA>