ConstructMetacells

This function takes a Seurat object and constructs averaged 'metacells' based on neighboring cells.

ConstructMetacells(
  seurat_obj,
  name = "agg",
  ident.group = "seurat_clusters",
  k = 25,
  reduction = "umap",
  assay = "RNA",
  cells.use = NULL,
  slot = "counts",
  meta = NULL,
  return_metacell = FALSE,
  mode = "average",
  max_shared = 15,
  target_metacells = 1000,
  max_iter = 5000,
  verbose = FALSE,
  wgcna_name = NULL
)

Arguments

seurat_obj

A Seurat object

name

A string appended to resulting metalcells. Default = 'agg'

k

Number of nearest neighbors to aggregate. Default = 50

reduction

A dimensionality reduction stored in the Seurat object. Default = 'umap'

assay

Assay to extract data for aggregation. Default = 'RNA'

slot

Slot to extract data for aggregation. Default = 'data'

return_metacell

Logical to determine if we return the metacell seurat object (TRUE), or add it to the misc in the original Seurat object (FALSE). Default to FALSE.

mode

determines how to make gene expression profiles for metacells from their constituent single cells. Options are "average" or "sum".

max_shared

the maximum number of cells to be shared across two metacells

target_metacells

the maximum target number of metacells to construct

max_iter

the maximum number of iterations in the metacells bootstrapping loop

verbose

logical indicating whether to print additional information

wgcna_name

name of the WGCNA experiment

Examples

ConstructMetacells
#> function (seurat_obj, name = "agg", ident.group = "seurat_clusters", 
#>     k = 25, reduction = "umap", assay = "RNA", cells.use = NULL, 
#>     slot = "counts", meta = NULL, return_metacell = FALSE, mode = "average", 
#>     max_shared = 15, target_metacells = 1000, max_iter = 5000, 
#>     verbose = FALSE, wgcna_name = NULL) 
#> {
#>     if (is.null(wgcna_name)) {
#>         wgcna_name <- seurat_obj@misc$active_wgcna
#>     }
#>     if (!(reduction %in% names(seurat_obj@reductions))) {
#>         stop(paste0("Invalid reduction (", reduction, "). Reductions in Seurat object: ", 
#>             paste(names(seurat_obj@reductions), collapse = ", ")))
#>     }
#>     if (!(assay %in% names(seurat_obj@assays))) {
#>         stop(paste0("Invalid assay (", assay, "). Assays in Seurat object: ", 
#>             paste(names(seurat_obj@assays), collapse = ", ")))
#>     }
#>     if (!(slot %in% c("counts", "data", "scale.data"))) {
#>         stop(paste0("Invalid slot (", slot, "). Valid options for slot: counts, data, scale.data "))
#>     }
#>     if (!is.null(cells.use)) {
#>         seurat_full <- seurat_obj
#>         seurat_obj <- seurat_obj[, cells.use]
#>     }
#>     reduced_coordinates <- as.data.frame(seurat_obj@reductions[[reduction]]@cell.embeddings)
#>     nn_map <- FNN::knn.index(reduced_coordinates, k = (k - 1))
#>     row.names(nn_map) <- row.names(reduced_coordinates)
#>     nn_map <- cbind(nn_map, seq_len(nrow(nn_map)))
#>     good_choices <- seq_len(nrow(nn_map))
#>     choice <- sample(seq_len(length(good_choices)), size = 1, 
#>         replace = FALSE)
#>     chosen <- good_choices[choice]
#>     good_choices <- good_choices[good_choices != good_choices[choice]]
#>     it <- 0
#>     k2 <- k * 2
#>     get_shared <- function(other, this_choice) {
#>         k2 - length(union(cell_sample[other, ], this_choice))
#>     }
#>     while (length(good_choices) > 0 & length(chosen) < target_metacells & 
#>         it < max_iter) {
#>         it <- it + 1
#>         choice <- sample(seq_len(length(good_choices)), size = 1, 
#>             replace = FALSE)
#>         new_chosen <- c(chosen, good_choices[choice])
#>         good_choices <- good_choices[good_choices != good_choices[choice]]
#>         cell_sample <- nn_map[new_chosen, ]
#>         others <- seq_len(nrow(cell_sample) - 1)
#>         this_choice <- cell_sample[nrow(cell_sample), ]
#>         shared <- sapply(others, get_shared, this_choice = this_choice)
#>         if (max(shared) <= max_shared) {
#>             chosen <- new_chosen
#>         }
#>     }
#>     shared_old <- shared
#>     cell_sample <- nn_map[chosen, ]
#>     combs <- tryCatch({
#>         combn(nrow(cell_sample), 2)
#>     }, error = function(cond) {
#>         return(NA)
#>     })
#>     if (any(is.na(combs))) {
#>         warning("Metacell failed")
#>         return(NULL)
#>     }
#>     shared <- apply(combs, 2, function(x) {
#>         k2 - length(unique(as.vector(cell_sample[x, ])))
#>     })
#>     if (verbose) {
#>         message(paste0("Overlap QC metrics:\nCells per bin: ", 
#>             k, "\nMaximum shared cells bin-bin: ", max(shared), 
#>             "\nMean shared cells bin-bin: ", mean(shared), "\nMedian shared cells bin-bin: ", 
#>             median(shared)))
#>         if (mean(shared)/k > 0.1) 
#>             warning("On average, more than 10% of cells are shared between paired bins.")
#>     }
#>     exprs_old <- GetAssayData(seurat_obj, assay = assay, slot = slot)
#>     mask <- sapply(seq_len(nrow(cell_sample)), function(x) seq_len(ncol(exprs_old)) %in% 
#>         cell_sample[x, , drop = FALSE])
#>     mask <- Matrix::Matrix(mask)
#>     new_exprs <- (exprs_old %*% mask)
#>     if (mode == "average") {
#>         new_exprs <- new_exprs/k
#>     }
#>     colnames(new_exprs) <- paste0(name, "_", 1:ncol(new_exprs))
#>     rownames(cell_sample) <- paste0(name, "_", 1:ncol(new_exprs))
#>     colnames(cell_sample) <- paste0("knn_", 1:ncol(cell_sample))
#>     metacell_obj <- CreateSeuratObject(counts = new_exprs, assay = assay)
#>     max_shared <- max(shared)
#>     median_shared <- median(shared)
#>     mean_shared <- mean(shared)
#>     new_exprs[new_exprs > 0] <- 1
#>     density <- sum(Matrix::colSums(new_exprs)/(nrow(new_exprs) * 
#>         ncol(new_exprs)))
#>     run_stats <- data.frame(name = name, max_shared = max_shared, 
#>         mean_shared = mean_shared, median_shared = median_shared, 
#>         density = density, n = ncol(new_exprs))
#>     metacell_obj@misc$run_stats <- run_stats
#>     if (!is.null(meta)) {
#>         meta_names <- names(meta)
#>         for (x in meta_names) {
#>             metacell_obj@meta.data[[x]] <- meta[[x]]
#>         }
#>     }
#>     else (warning("meta not found"))
#>     if (return_metacell) {
#>         out <- metacell_obj
#>     }
#>     else {
#>         if (!is.null(cells.use)) {
#>             seurat_obj <- seurat_full
#>         }
#>         seurat_obj <- SetMetacellObject(seurat_obj, metacell_obj, 
#>             wgcna_name)
#>         seurat_obj <- SetWGCNAParams(seurat_obj, params = list(metacell_k = k, 
#>             metacell_reduction = reduction, metacell_slot = slot, 
#>             metacell_assay = assay), wgcna_name)
#>         out <- seurat_obj
#>     }
#>     out
#> }
#> <bytecode: 0x7fb02cc0bf50>
#> <environment: namespace:hdWGCNA>