SetMultiExpr

This function sets up the expression matrix input for consensus WGCNA based on the metacell expression matrix, the full expression matrix, or a provided pseudobulk expression matrix.

Usage

SetMultiExpr(
  seurat_obj,
  group_name,
  use_metacells = TRUE,
  group.by = NULL,
  multi.group.by = NULL,
  multi_groups = NULL,
  assay = NULL,
  slot = "data",
  layer = "data",
  mat = NULL,
  mat_group_delim = 3,
  wgcna_name = NULL,
  ...
)

Arguments

seurat_obj

A Seurat object

group_name

A string containing a group present in the provided group.by column or in the Seurat Idents.

use_metacells

A logical determining if we use the metacells (TRUE) or the full expression matrix (FALSE)

group.by

A string containing the name of a column in the Seurat object with cell groups (clusters, cell types, etc). If NULL (default), hdWGCNA uses the Seurat Idents as the group.

multi.group.by

A string containing the name of a column in the Seurat object with groups for consensus WGCNA (dataset, sample, condition, etc)

multi_groups

A character vecrtor containing the names of groups to select

assay

The name of the assay in the Seurat object

slot

The name of the slot in the Seurat object (counts, data)

layer

Layer to extract data for aggregation. Default = 'counts'. Layer is used with Seurat v5 instead of slot.

mat

A Matrix containing gene expression data. Supplying a matrix using this parameter ignores other options. This is almost exclusively used for pseudobulk analysis.

wgcna_name

A string containing the name of the WGCNA slot in seurat_obj@misc. Default = NULL which retrieves the currently active WGCNA data