SetMultiExpr

This function sets up the expression matrix input for consensus WGCNA based on the metacell expression matrix, or from the full expression matrix.

SetMultiExpr(
  seurat_obj,
  group_name,
  use_metacells = TRUE,
  group.by = NULL,
  multi.group.by = NULL,
  multi_groups = NULL,
  wgcna_name = NULL,
  assay = NULL,
  slot = "data",
  ...
)

Arguments

seurat_obj

A Seurat object

group_name

A string containing a group present in the provided group.by column or in the Seurat Idents.

use_metacells

A logical determining if we use the metacells (TRUE) or the full expression matrix (FALSE)

group.by

A string containing the name of a column in the Seurat object with cell groups (clusters, cell types, etc). If NULL (default), hdWGCNA uses the Seurat Idents as the group.

multi.group.by

A string containing the name of a column in the Seurat object with groups for consensus WGCNA (dataset, sample, condition, etc)

multi_groups

A character vecrtor containing the names of groups to select

wgcna_name

A string containing the name of the WGCNA slot in seurat_obj@misc. Default = NULL which retrieves the currently active WGCNA data

Examples

SetDatExpr
#> function (seurat_obj, group_name, use_metacells = TRUE, group.by = NULL, 
#>     multi.group.by = NULL, multi_group_name = NULL, return_seurat = TRUE, 
#>     wgcna_name = NULL, assay = NULL, slot = "data", ...) 
#> {
#>     if (is.null(wgcna_name)) {
#>         wgcna_name <- seurat_obj@misc$active_wgcna
#>     }
#>     params <- GetWGCNAParams(seurat_obj, wgcna_name)
#>     genes_use <- GetWGCNAGenes(seurat_obj, wgcna_name)
#>     modules <- GetModules(seurat_obj, wgcna_name)
#>     m_obj <- GetMetacellObject(seurat_obj, wgcna_name)
#>     if (use_metacells & !is.null(m_obj)) {
#>         s_obj <- m_obj
#>     }
#>     else {
#>         if (is.null(m_obj)) {
#>             warning("Metacell Seurat object not found. Using full Seurat object instead.")
#>         }
#>         s_obj <- seurat_obj
#>     }
#>     seurat_meta <- s_obj@meta.data
#>     if (is.null(assay)) {
#>         assay <- DefaultAssay(s_obj)
#>         warning(paste0("assay not specified, trying to use assay ", 
#>             assay))
#>     }
#>     if (!(assay %in% names(s_obj@assays))) {
#>         stop("Assay not found. Check names(seurat_obj@assays) or names(GetMetacellObject(seurat_obj)@assays)")
#>     }
#>     if (!is.null(group.by)) {
#>         if (!(group.by %in% colnames(s_obj@meta.data))) {
#>             m_cell_message <- ""
#>             if (use_metacells) {
#>                 m_cell_message <- "metacell"
#>             }
#>             stop(paste0(group.by, " not found in the meta data of the ", 
#>                 m_cell_message, " Seurat object"))
#>         }
#>     }
#>     if (!is.null(group.by)) {
#>         seurat_meta <- seurat_meta %>% subset(get(group.by) %in% 
#>             group_name)
#>     }
#>     if (!is.null(multi.group.by)) {
#>         seurat_meta <- seurat_meta %>% subset(get(multi.group.by) %in% 
#>             multi_group_name)
#>     }
#>     cells <- rownames(seurat_meta)
#>     datExpr <- as.data.frame(Seurat::GetAssayData(s_obj, assay = assay, 
#>         slot = slot)[genes_use, cells])
#>     datExpr <- as.data.frame(t(datExpr))
#>     if (is.null(multi.group.by)) {
#>         gene_list = genes_use[WGCNA::goodGenes(datExpr, ...)]
#>         datExpr <- datExpr[, gene_list]
#>     }
#>     if (return_seurat) {
#>         seurat_obj <- SetWGCNAGenes(seurat_obj, gene_list, wgcna_name)
#>         seurat_obj@misc[[wgcna_name]]$datExpr <- datExpr
#>         out <- seurat_obj
#>     }
#>     else {
#>         out <- datExpr
#>     }
#>     out
#> }
#> <bytecode: 0x7ff68644b3c8>
#> <environment: namespace:hdWGCNA>